finished initialization of the slab allocator and tweaked some object sizes to be a multiple of 4 bytes

+++ kernel/thread.h
@@ -8,7 +8,7 @@ struct Thread;
 struct Thread {
 	uint32_t regs[17];
 	uint16_t tid;
-	uint8_t status;
+	uint16_t status;
 };
